left component bets persistance works

+++ db/migrate/20200423192146_create_games.rb
@@ -1,7 +1,7 @@
 class CreateGames < ActiveRecord::Migration[6.0]
   def change
     create_table :games do |t|
-      t.integer :jackpot
+      t.float :jackpot
       t.integer :minimum_bet
       t.string :winner
       t.integer :bets_placed, :default => 0
